api/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
pydantic>=2.0.0
tiktoken>=0.5.0
adalflow>=0.1.0
numpy>=1.24.0
faiss-cpu>=1.7.4
langid>=1.1.6
requests>=2.28.0
jinja2>=3.1.2
python-dotenv>=1.0.0
openai>=1.76.2
aiohttp>=3.8.4
websockets>=11.0.3
backoff>=2.2.0

